+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-*March 3rd, 2022*
+**March 3rd, 2022**
 - (DCh) added  background placeholder for the starting loc and the model of the main charachter
+
+**March 12th, 2022**
+- (DCh) added character movement
+- (DCh) camera follows character